pyproject.toml
[tool.poetry]
name = "report-deployer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
PyYAML = "^6.0.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
